NXP-23481: extract blob zip utils
